+++ front-end/src/router/routes.js
@@ -5,14 +5,8 @@ const routes = [
     component: () => import('layouts/MainLayout.vue'),
     children: [
       { path: '',  name: 'Index', component: () => import('pages/IndexPage.vue') },
-      {
-        path: '/ticket',
-        name: 'Tickets',
-        component: () => import('pages/ticket/TicketIndexPage.vue'),
-        props: true,
-        children: [
-          { path: '/ticket/:id', name: 'TicketEditor', component: () => import('pages/ticket/TicketPage.vue'), props: true },
-      ]},
+      { path: '/ticket', name: 'TicketIndex', component: () => import('pages/ticket/TicketIndexPage.vue') },
+      { path: '/ticket/:id', name: 'TicketEditor', component: () => import('pages/ticket/TicketPage.vue'), props: true },
       { path: '/article',  name: 'Articles', component: () => import('pages/ArticlePage.vue') }
     ]
   },
